Cleaning up the Reed-Solomon decoder tests.
Rewrote the source for the decoder tests in the same format at the
source for the encoder and decoder (i.e. header files that are included
into the bodies of other funtions). This unifies the way the
Reed-Solomon library is written, and reduces code duplication.

This change also makes it easier to extend the tests.