Adding more tests for the RS encoder/decoder.
The tests for the encoder/decoder had the following deficits:

The decoder should be able to correct any error and erasure pattern as
long as 2 * #errors + #erasures < minimun_distance. The previous version
of the tests was restricted to error and erasure patterns with #errors
+ #erasures < minimun_distance / 2.

The decoder should either report failure or return a valid codeword.
When testing up to error correction capacity, the check for this is
automatic. However, there were no tests to ensure that the decoder
behaves as it should when we exceed the error correction capacity.
To be more precise; when we exceed the error correction capacity there
is no guaratee that the decoder decodes correctly. It should, however,
always return a valid codeword, unless it specifically reports failure.

Only full length codes (i.e. pad = 0) were used for the tests.

This commit fixes these issues.