Updating RS encoder/decoder tests.
The new tests (up to error correction capacity) now checks that decoder
returns the number of symbols actually corrected. This means that
erasures where the symbol isn't corrupted are not included in the count
of corrected symbols. Also, the returned error positions only include
positions where a symbol was actually fixed.