Fixing bugs in the RS decoder.
Make sure that the decoder actually returns a codeword, or announces a
decoding failure. We also introduce new return codes for the decoder.
These return codes distinguish between the different 'causes' for an
uncorrectable error. These error codes are probably not useful to users
of the library, so they are for internal use also.

closes #2